+++ apps/backend/src/app.module.ts
@@ -2,12 +2,20 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
+import { AuthModule } from './auth/auth.module';
+import { FamiliesModule } from './families/families.module';
+import { PrismaModule } from './prisma/prisma.module';
+import { UsersModule } from './users/users.module';
 
 @Module({
   imports: [
     ConfigModule.forRoot({
       isGlobal: true
-    })
+    }),
+    PrismaModule,
+    AuthModule,
+    FamiliesModule,
+    UsersModule
   ],
   controllers: [AppController],
   providers: [AppService]